Simplify AlbumsService implementation
