+++ go.mod
@@ -3,19 +3,10 @@ module github.com/gphotosuploader/google-photos-api-client-go/v2
 go 1.18
 
 require (
-	github.com/gadelkareem/cachita v0.2.3
 	github.com/gorilla/mux v1.8.0
 	github.com/gphotosuploader/googlemirror v0.5.0
 	github.com/hashicorp/go-retryablehttp v0.7.4
 	google.golang.org/api v0.126.0
 )
 
-require (
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
-	github.com/mediocregopher/radix/v3 v3.2.0 // indirect
-	github.com/vmihailenco/msgpack v4.0.1+incompatible // indirect
-	golang.org/x/net v0.10.0 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.30.0 // indirect
-)
+require github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
